combine proUtil and proReHandlers
